Backend/requirements.txt
# Web framework
Flask==2.3.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1

# Database
mysql-connector-python==8.1.0
SQLAlchemy==2.0.25
PyMySQL==1.1.0

# PDF / DOCX parsing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==0.8.11

# NLP / AI / ML
spacy==3.7.1
nltk==3.8.1
transformers==4.34.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.26.0
pandas==2.1.1
torch==2.1.0

# API / Requests
openai==1.7.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
bcrypt==4.1.0
email-validator==2.0.0
passlib==1.7.4
